Keep libjseval.c contract-compatible; move sandbox additions to wasmlib.c
libjseval.a is also linked into the NEAR contract (build.rs), so the new
js_host_time_ms import broke contract wasm instantiation and the js_eval
JSValue return type mismatched the i32 declaration in src/jslib.rs.
Revert libjseval.c to the original and implement the full-JSValue eval,
interrupt handler and memory limit in wasmlib.c, which is only compiled
into the npm package's jseval.wasm.

build.sh again builds libquickjs.a in the repo-root QuickJS checkout when
present (build.rs links it from there) and sources ../emsdk/emsdk_env.sh
when emcc is not on the PATH, falling back to the pinned hermetic
download for standalone/npm CI builds.
