+++ quickjslib/build.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
-# Reproducible build of jseval.wasm.
+# Build of jseval.wasm, reproducible per toolchain.
 #
-# Pins the QuickJS version (downloaded from bellard.org if not present) and,
-# when emcc is not already on the PATH, bootstraps a pinned emsdk. The
-# QuickJS source is extracted into this directory (gitignored) so the build
-# does not depend on — or interfere with — checkouts elsewhere in the repo.
+# This script serves two contexts:
+# - The repo build (invoked by ../build.rs): uses the QuickJS checkout and
+#   emsdk at the repo root. build.rs links libquickjs.a from that QuickJS
+#   directory, so it must be built there.
+# - Standalone (npm package CI / fresh clone): downloads the pinned QuickJS
+#   release into this directory (gitignored) and uses emcc from the PATH,
+#   bootstrapping a pinned emsdk if none is installed.
 set -euo pipefail
 
 QUICKJS_VERSION=2026-06-04
@@ -13,12 +16,21 @@ EMSDK_VERSION=3.1.74
 SCRIPT_DIR=$(cd "$(dirname "$0")" && pwd)
 cd "$SCRIPT_DIR"
 
-QUICKJS_ROOT="$SCRIPT_DIR/quickjs-$QUICKJS_VERSION"
+if [ -d "../quickjs-$QUICKJS_VERSION" ]; then
+    QUICKJS_ROOT=$(cd "../quickjs-$QUICKJS_VERSION" && pwd)
+else
+    QUICKJS_ROOT="$SCRIPT_DIR/quickjs-$QUICKJS_VERSION"
+    if [ ! -d "$QUICKJS_ROOT" ]; then
+        curl -fLO "https://bellard.org/quickjs/quickjs-$QUICKJS_VERSION.tar.xz"
+        tar -xf "quickjs-$QUICKJS_VERSION.tar.xz"
+        rm "quickjs-$QUICKJS_VERSION.tar.xz"
+    fi
+fi
 
-if [ ! -d "$QUICKJS_ROOT" ]; then
-    curl -fLO "https://bellard.org/quickjs/quickjs-$QUICKJS_VERSION.tar.xz"
-    tar -xf "quickjs-$QUICKJS_VERSION.tar.xz"
-    rm "quickjs-$QUICKJS_VERSION.tar.xz"
+if ! command -v emcc > /dev/null && [ -f ../emsdk/emsdk_env.sh ]; then
+    set +u
+    source ../emsdk/emsdk_env.sh || true
+    set -u
 fi
 
 if ! command -v emcc > /dev/null; then
@@ -27,7 +39,9 @@ if ! command -v emcc > /dev/null; then
     fi
     (cd "$SCRIPT_DIR/emsdk" && git fetch --tags && git checkout "$EMSDK_VERSION" \
         && ./emsdk install "$EMSDK_VERSION" && ./emsdk activate "$EMSDK_VERSION")
+    set +u
     source "$SCRIPT_DIR/emsdk/emsdk_env.sh"
+    set -u
 fi
 
 echo "Building with $(emcc --version | head -1)"
